requirements.txt
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
djangorestframework-recursive
django-bootstrap3
requests